pyproject.toml
[project]
name = "ais-analysis"
version = "3.0.0"
description = "High-performance S3-based pipeline for processing AIS data from the Danish Maritime Authority"
authors = [
    {name = "Paul Singerman", email = "saulpingerman@users.noreply.github.com"}
]
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Topic :: Scientific/Engineering :: GIS",
]
keywords = ["ais", "maritime", "data-processing", "s3", "aws", "analytics", "vessel-tracking"]

dependencies = [
    "polars>=1.30.0",
    "pyarrow>=20.0.0",
    "pandas>=2.0.0",
    "boto3",
    "PyYAML>=6.0",
    "geopy>=2.4.0",
    "tqdm>=4.60.0",
    "tabulate",
    "requests",
    "scikit-learn>=1.0.0",
    "pydantic>=2.0.0",
    "typer>=0.9.0",
    "s3fs>=2023.0.0",
    "fsspec>=2023.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]
viz = [
    "matplotlib>=3.9.0",
]

[project.urls]
Homepage = "https://github.com/saulpingerman/ais-analysis"
Repository = "https://github.com/saulpingerman/ais-analysis"
Issues = "https://github.com/saulpingerman/ais-analysis/issues"

[project.scripts]
ais-download = "scripts.download_ais_data:main"
ais-process-legacy = "scripts.s3_ais_processor:main"
ais-pipeline = "ais_pipeline.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ais_pipeline", "scripts", "tests"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/scripts",
    "/tests",
    "/config",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
